rename: endpoint name 수정
