+++ src/user/user.service.ts
@@ -28,4 +28,15 @@ export class UserService {
     const createdUser = await this.userRepository.findUserById(id);
     return { message: `${createdUser.nickname}에 대한 최초 정보 기입 성공.` };
   }
+
+  //회원탈퇴
+  async deleteUser(id: number) {
+    const user = await this.userRepository.findUserById(id);
+    if (!user) {
+      throw new BadRequestException(
+        '해당 계정이 존재하지 않아 삭제할 수 없습니다.',
+      );
+    }
+    await this.userRepository.softDeleteUser(id);
+  }
 }
